more Exception flow refactor
