+++ athena-scripts/raw_stage_optimisation_solution/scripts/raw_to_stage_etl/exceptions/NoDataFoundException.py
