+++ athena-scripts/raw_stage_optimisation_solution/scripts/raw_to_stage_etl/strategies/Strategy.py
@@ -7,12 +7,12 @@ from datetime import datetime
 
 import pandas as pd
 
+from ..exceptions.NoDataFoundException import NoDataFoundException
 from ..logger import logger
 from ..util.processing_utilities import (add_new_column, add_new_column_from_struct, empty_string_to_null,
                                          extract_element_by_name_and_validate, generate_key_value_records,
                                          remove_columns, remove_row_duplicates, remove_rows_missing_mandatory_values,
                                          rename_column_names)
-from .exceptions.NoDataFoundException import NoDataFoundException
 
 
 class Strategy(ABC):
