+++ athena-scripts/raw_stage_optimisation_solution/scripts/setup.cfg
@@ -12,9 +12,9 @@ packages = find:
 include =
     raw_to_stage_etl
     raw_to_stage_etl.clients
+    raw_to_stage_etl.exceptions
+    raw_to_stage_etl.logger
     raw_to_stage_etl.processor
     raw_to_stage_etl.strategies
-    raw_to_stage_etl.strategies.exceptions
     raw_to_stage_etl.util
     raw_to_stage_etl.util.exceptions
-    raw_to_stage_etl.logger
